requirements.txt
# Core Dependencies
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0
scipy>=1.7.0,<2.0.0

# Deep Learning
torch>=2.0.0,<3.0.0
torch-geometric>=2.3.0,<3.0.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# ROS2 (Humble)
rclpy>=3.0.0

# Visualization
matplotlib>=3.5.0,<4.0.0
plotly>=5.0.0,<6.0.0
seaborn>=0.11.0,<1.0.0

# Data Processing
pyyaml>=6.0,<7.0
tqdm>=4.62.0,<5.0.0
scikit-learn>=1.0.0,<2.0.0

# Utilities
python-dateutil>=2.8.0
